broke up big file
